refactor: 독후감 작성 화면 탭 레이아웃으로 변경
